Code cleanup
Tested with shoelace following. Though it is not the most robust piece of software out there,
it worked. The lines were detected nicely, and the reduced frame rate helped the overall system.
The final output was horizontal distance to center of the frame, which was printed to terminal
only to be piped to the control code.