Added the red ball tracking script
Based on: https://www.pyimagesearch.com/2015/09/14/ball-tracking-with-opencv/
Hardcoded to track a red-ish House M.D. ball. redLower and redUpper can be modified to different HSV ranges.

A color-mask is implemented and we avoid using HoughCircles and instead use contourArea, minEnclosingCircle and moments functions in opencv. We return the horizontal distance from center of the frame.